- Fix build break (GFX buffer manager working in progress) - Fix display mode auto-detection bug on multi monitor system - Fix various vc8 build breaks.

+++ src/priv/gfx/d3d/d3dBufferMgr.cpp
@@ -58,7 +58,7 @@ void GN::gfx::D3DRenderer::bindVtxBufs( const VtxBuf * const buffers[], uint32_t
                 stage,
                 safeCast<const D3DVtxBuf*>(buffers[i])->getD3DVb(),
                 0,
-                buffers[i]->getStride() ) );
+                (UINT)buffers[i]->getStride() ) );
         }
         else
         {
